backend/requirements.txt
fastapi
uvicorn
python-multipart
llama-index
llama-index-vector-stores-chroma
google-genai
chromadb
pypdf
python-dotenv
pydantic
pinecone-client